Fetch master ref on build
This allows a project to change its build strategy depending on whether
we're building master or a branch.

A concrete use case for this is: on public-web, we check in the
translation manifest on master, but not on branches (to avoid nasty
merges). We have a validation script that checks the manifest, and
currently it will fail on branches that add new translatable strings.

This patch would allow public-web to run different manifest validations
depending on whether we are building master or a branch.