Remove setsid from bin/launch script
setsid is no longer needed after a change to p2. In addition this is now
causing bad behavior... runit is spawning hundreds of kochiku-worker
processes.

Explanation from xunil: the setsid command will fork() for you if you
are already the process group leader, because setsid(2) fails if you are
process group leader